set up two clocks in DT "sys_clk" and "apb_clk" where the first is registered in drivers/clk/clk-moxart.c and the second is "fixed-clock".
this means the "PMU" register can be mapped once in clock driver and MMC (or future drivers) can then request the clock with devm_clk_get.

* remove clock read from MMC
* remove reg_pmu from MMC (and from DT)
* add of_ debug pr_info in drivers/clk/clk.c drivers/clk/clkdev.c (remove when finished)

+++ drivers/mmc/host/sdhci-moxart.h
@@ -126,7 +126,7 @@ struct moxart_host {
 
 	unsigned int			wait_for;
 	unsigned int			timeout;
-	unsigned int			sysclk;
+	long					sysclk;
 	bool					removed;
 };
